requirements.txt
openai>=1.0.0
 pypdf>=3.14.0
python-dotenv>=0.21.0
rich>=13.0.0
pytest>=7.0.0
Pillow>=10.0.0
pytesseract>=0.3.10